final/app/requirements.txt
streamlit>=1.31
torch>=2.1
torchvision>=0.16
Pillow>=10.0
